+++ CHANGES.rst
@@ -11,6 +11,9 @@ New Features:
 
 Bugfixes:
 
+- Fix the use of fullobjects in Archetypes based sites @search
+  [erral]
+
 - Fix workflow translations with unicode characters.
   [Gagaro]